+++ fdb-record-layer-lucene/src/test/java/com/apple/foundationdb/record/lucene/LucenePrimaryKeySegmentIndexTest.java
@@ -286,10 +286,13 @@ public class LucenePrimaryKeySegmentIndexTest extends FDBRecordStoreTestBase {
             final LuceneIndexMaintainer indexMaintainer = (LuceneIndexMaintainer)recordStore.getIndexMaintainer(index);
             // TODO improve this as part of #2575
             final FailCommitsAgilityContext agilityContext = new FailCommitsAgilityContext(context, index.getSubspaceKey());
-            assertThrows(FailedLuceneCommit.class,
-                    () -> indexMaintainer.mergeIndexForTesting(Tuple.from(), null, agilityContext));
+            try {
+                assertThrows(FailedLuceneCommit.class,
+                        () -> indexMaintainer.mergeIndexForTesting(Tuple.from(), null, agilityContext));
+            } finally {
+                agilityContext.flushAndClose();
+            }
             assertEquals(1, agilityContext.commitCount);
-            agilityContext.flushAndClose(); // the normal merge would handle this
         }
         // V1 fails here, that's the test
         Assumptions.assumeTrue(version == Version.V2);
@@ -316,11 +319,6 @@ public class LucenePrimaryKeySegmentIndexTest extends FDBRecordStoreTestBase {
             rebuildIndexMetaData(context, SIMPLE_DOC, index);
             final LuceneIndexMaintainer indexMaintainer = (LuceneIndexMaintainer)recordStore.getIndexMaintainer(index);
             indexMaintainer.mergeIndex();
-            // This is required to close the agility context
-            // the pre-commit hook on this transaction will close the DirectoryManager
-            // which will close the directory and flush the context
-            // See also: #2574
-            context.commit();
         }
         try (FDBRecordContext context = openContext()) {
             rebuildIndexMetaData(context, SIMPLE_DOC, index);