+++ fdb-record-layer-lucene/src/test/java/com/apple/foundationdb/record/lucene/codec/LuceneOptimizedDocValuesFormatTest.java
@@ -49,6 +49,7 @@ import org.apache.lucene.util.BytesRef;
 import org.apache.lucene.util.TestRuleLimitSysouts;
 import org.apache.lucene.util.TestUtil;
 import org.junit.BeforeClass;
+import org.junit.Ignore;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -371,5 +372,19 @@ public class LuceneOptimizedDocValuesFormatTest extends BaseDocValuesFormatTestC
         ir.close();
         writer.close();
         dir.close();
+    }
+
+    @Override
+    @Ignore // Issue #2598: Make Lucene @Nightly tests pass for fixed seed
+    @Nightly
+    public void testRamBytesUsed() throws IOException {
+        super.testRamBytesUsed();
+    }
+
+    @Override
+    @Ignore // Issue #2598: Make Lucene @Nightly tests pass for fixed seed
+    @Nightly
+    public void testThreads3() throws Exception {
+        super.testThreads3();
     }
 }